squawk/requirements.txt
flask
pytest
psycopg2-binary

